+++ tests/Choice/ContentTypeTest.php
@@ -14,7 +14,7 @@ class ContentTypeTest extends AbstractBoltFormsUnitTest
     public function testGetName()
     {
         $app = $this->getApp();
-        $obj = new ContentType($app, 'koala', 'contenttype::pages::title::slug');
+        $obj = new ContentType($app, 'koala', array('choices' => 'contenttype::pets::title::slug'));
         $this->assertInstanceOf('\Bolt\Extension\Bolt\BoltForms\Choice\ContentType', $obj);
         $this->assertSame($obj->getName(), 'koala');
     }
@@ -25,11 +25,11 @@ class ContentTypeTest extends AbstractBoltFormsUnitTest
     public function testGetChoicesException()
     {
         $app = $this->getApp();
-        $obj = new ContentType($app, 'koala', 'contenttype');
+        $obj = new ContentType($app['storage'], 'koala', array('choices' => 'contenttype'));
         $obj->getChoices();
-        $obj = new ContentType($app, 'koala', 'contenttype::pages');
+        $obj = new ContentType($app['storage'], 'koala', array('choices' => 'contenttype::pages'));
         $obj->getChoices();
-        $obj = new ContentType($app, 'koala', 'contenttype::pages::title');
+        $obj = new ContentType($app['storage'], 'koala', array('choices' => 'contenttype::pages::title'));
         $obj->getChoices();
     }
 
@@ -47,7 +47,7 @@ class ContentTypeTest extends AbstractBoltFormsUnitTest
             ->will($this->returnValue(array($record)));
         $app['storage'] = $storage;
 
-        $obj = new ContentType($app, 'koala', 'contenttype::pages::title::slug');
+        $obj = new ContentType($app['storage'], 'koala', array('choices' => 'contenttype::pets::title::slug'));
         $choice = $obj->getChoices();
 
         $this->assertNotEmpty($choice);